refactor: split api session runtime and stream helpers
