+++ internal/api/handler_session_chat.go
@@ -1,7 +1,6 @@
 package api
 
 import (
-	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -11,16 +10,12 @@ import (
 	"os/exec"
 	"strconv"
 	"strings"
-	"time"
 
-	"github.com/danielgtaylor/huma/v2/sse"
 	"github.com/gastownhall/gascity/internal/beads"
 	"github.com/gastownhall/gascity/internal/config"
 	"github.com/gastownhall/gascity/internal/runtime"
 	"github.com/gastownhall/gascity/internal/session"
-	"github.com/gastownhall/gascity/internal/sessionlog"
 	"github.com/gastownhall/gascity/internal/shellquote"
-	workdirutil "github.com/gastownhall/gascity/internal/workdir"
 	"github.com/gastownhall/gascity/internal/worker"
 )
 
@@ -71,191 +66,6 @@ type sessionRawTranscriptResponse struct {
 	Format     string                       `json:"format"`
 	Messages   []json.RawMessage            `json:"messages"`
 	Pagination *worker.TranscriptPagination `json:"pagination,omitempty"`
-}
-
-// SessionStreamMessageEvent carries normalized conversation turns on the
-// session SSE stream.
-type SessionStreamMessageEvent struct {
-	ID         string                     `json:"id"`
-	Template   string                     `json:"template"`
-	Provider   string                     `json:"provider" doc:"Producing provider identifier (claude, codex, gemini, open-code, etc.)."`
-	Format     string                     `json:"format"`
-	Turns      []outputTurn               `json:"turns"`
-	Pagination *sessionlog.PaginationInfo `json:"pagination,omitempty"`
-}
-
-// SessionStreamRawMessageEvent carries provider-native transcript frames on
-// the session SSE stream.
-type SessionStreamRawMessageEvent struct {
-	ID         string                     `json:"id"`
-	Template   string                     `json:"template"`
-	Provider   string                     `json:"provider" doc:"Producing provider identifier (claude, codex, gemini, open-code, etc.). Consumers use this to dispatch per-provider frame parsing."`
-	Format     string                     `json:"format"`
-	Messages   []SessionRawMessageFrame   `json:"messages" doc:"Provider-native transcript frames, emitted verbatim as the provider wrote them."`
-	Pagination *sessionlog.PaginationInfo `json:"pagination,omitempty"`
-}
-
-func (s *Server) sessionLogPaths() []string {
-	if s.sessionLogSearchPaths != nil {
-		return s.sessionLogSearchPaths
-	}
-	cfg := s.state.Config()
-	if cfg == nil {
-		return worker.DefaultSearchPaths()
-	}
-	return worker.MergeSearchPaths(cfg.Daemon.ObservePaths)
-}
-
-func sessionCreateHints(resolved *config.ResolvedProvider) runtime.Config {
-	return runtime.Config{
-		ReadyPromptPrefix:      resolved.ReadyPromptPrefix,
-		ReadyDelayMs:           resolved.ReadyDelayMs,
-		ProcessNames:           resolved.ProcessNames,
-		EmitsPermissionWarning: resolved.EmitsPermissionWarning,
-	}
-}
-
-func sessionResumeHints(resolved *config.ResolvedProvider, workDir string) runtime.Config {
-	return runtime.Config{
-		WorkDir:                workDir,
-		ReadyPromptPrefix:      resolved.ReadyPromptPrefix,
-		ReadyDelayMs:           resolved.ReadyDelayMs,
-		ProcessNames:           resolved.ProcessNames,
-		EmitsPermissionWarning: resolved.EmitsPermissionWarning,
-		Env:                    resolved.Env,
-	}
-}
-
-func sessionExplicitNameForCreate(agentCfg config.Agent, alias string) (string, error) {
-	if !agentCfg.SupportsMultipleSessions() || strings.TrimSpace(alias) != "" {
-		return "", nil
-	}
-	return session.GenerateAdhocExplicitName(agentCfg.Name)
-}
-
-func (s *Server) resolveSessionWorkDir(agentCfg config.Agent, qualifiedName string) (string, error) {
-	cfg := s.state.Config()
-	if cfg == nil {
-		return "", errors.New("no city config loaded")
-	}
-	workDir, err := workdirutil.ResolveWorkDirPathStrict(
-		s.state.CityPath(),
-		workdirutil.CityName(s.state.CityPath(), cfg),
-		qualifiedName,
-		agentCfg,
-		cfg.Rigs,
-	)
-	if err != nil {
-		return "", err
-	}
-	if workDir == "" {
-		workDir = s.state.CityPath()
-	}
-	return workDir, nil
-}
-
-// resolveSessionTemplateWithBareNameFallback resolves a session template
-// by name, retrying with the qualified name when the input is a bare
-// agent name that matches exactly one configured agent. Keeps the
-// two-phase lookup out of the handler.
-func (s *Server) resolveSessionTemplateWithBareNameFallback(name string) (*config.ResolvedProvider, string, string, string, error) {
-	resolved, workDir, transport, template, err := s.resolveSessionTemplate(name)
-	if err == nil {
-		return resolved, workDir, transport, template, nil
-	}
-	if !errors.Is(err, errSessionTemplateNotFound) || strings.Contains(name, "/") {
-		return nil, "", "", "", err
-	}
-	agentCfg, ok := findUniqueAgentTemplateByBareName(s.state.Config(), name)
-	if !ok {
-		return nil, "", "", "", err
-	}
-	return s.resolveSessionTemplate(agentCfg.QualifiedName())
-}
-
-func (s *Server) resolveSessionTemplate(template string) (*config.ResolvedProvider, string, string, string, error) {
-	cfg := s.state.Config()
-	if cfg == nil {
-		return nil, "", "", "", errors.New("no city config loaded")
-	}
-	agentCfg, ok := resolveSessionTemplateAgent(cfg, template)
-	if !ok {
-		return nil, "", "", "", errSessionTemplateNotFound
-	}
-	resolved, err := config.ResolveProvider(&agentCfg, &cfg.Workspace, cfg.Providers, exec.LookPath)
-	if err != nil {
-		return nil, "", "", "", err
-	}
-	workDir, err := s.resolveSessionWorkDir(agentCfg, agentCfg.QualifiedName())
-	if err != nil {
-		return nil, "", "", "", err
-	}
-	return resolved, workDir, agentCfg.Session, agentCfg.QualifiedName(), nil
-}
-
-func (s *Server) buildSessionResume(info session.Info) (string, runtime.Config) {
-	cmd := session.BuildResumeCommand(info)
-	resolved, workDir := s.resolveSessionRuntime(info)
-	if resolved == nil {
-		return cmd, runtime.Config{WorkDir: info.WorkDir}
-	}
-	resolvedInfo := info
-	resolvedInfo.Command = resolved.CommandString()
-	resolvedInfo.Provider = resolved.Name
-	resolvedInfo.ResumeFlag = resolved.ResumeFlag
-	resolvedInfo.ResumeStyle = resolved.ResumeStyle
-	resolvedInfo.ResumeCommand = resolved.ResumeCommand
-	return session.BuildResumeCommand(resolvedInfo), sessionResumeHints(resolved, workDir)
-}
-
-func (s *Server) resolveSessionRuntime(info session.Info) (*config.ResolvedProvider, string) {
-	kind := s.sessionKind(info.ID)
-	if kind != "provider" {
-		resolved, workDir, _, _, err := s.resolveSessionTemplate(info.Template)
-		if err == nil {
-			if info.WorkDir != "" {
-				workDir = info.WorkDir
-			}
-			return resolved, workDir
-		}
-	}
-
-	resolved, err := s.resolveBareProvider(info.Template)
-	if err != nil {
-		return nil, ""
-	}
-	workDir := info.WorkDir
-	if workDir == "" {
-		workDir = s.state.CityPath()
-	}
-	return resolved, workDir
-}
-
-// sessionKind reads the persisted mc_session_kind from bead metadata.
-func (s *Server) sessionKind(sessionID string) string {
-	store := s.state.CityBeadStore()
-	if store == nil {
-		return ""
-	}
-	b, err := store.Get(sessionID)
-	if err != nil {
-		return ""
-	}
-	return b.Metadata["mc_session_kind"]
-}
-
-// resolveBareProvider resolves a provider by name without an agent template.
-func (s *Server) resolveBareProvider(providerName string) (*config.ResolvedProvider, error) {
-	cfg := s.state.Config()
-	if cfg == nil {
-		return nil, errors.New("no city config loaded")
-	}
-	return config.ResolveProvider(
-		&config.Agent{Provider: providerName},
-		&cfg.Workspace,
-		cfg.Providers,
-		exec.LookPath,
-	)
 }
 
 func writeSessionManagerError(w http.ResponseWriter, err error) {
@@ -985,805 +795,4 @@ func (s *Server) handleSessionRespond(w http.ResponseWriter, r *http.Request) {
 	resp := map[string]string{"status": "accepted", "id": id}
 	s.idem.storeResponse(idemKey, bodyHash, http.StatusAccepted, resp)
 	writeJSON(w, http.StatusAccepted, resp)
-}
-
-func (s *Server) handleSessionStream(w http.ResponseWriter, r *http.Request) {
-	store := s.state.CityBeadStore()
-	if store == nil {
-		writeError(w, http.StatusServiceUnavailable, "unavailable", "no bead store configured")
-		return
-	}
-
-	id, err := s.resolveSessionIDAllowClosedWithConfig(store, r.PathValue("id"))
-	if err != nil {
-		writeResolveError(w, err)
-		return
-	}
-
-	catalog, err := s.workerSessionCatalog(store)
-	if err != nil {
-		writeSessionManagerError(w, err)
-		return
-	}
-	info, err := catalog.Get(id)
-	if err != nil {
-		writeSessionManagerError(w, err)
-		return
-	}
-	handle, err := s.workerHandleForSession(store, id)
-	if err != nil {
-		writeSessionManagerError(w, err)
-		return
-	}
-	historyReq := worker.HistoryRequest{}
-	if r.URL.Query().Get("format") == "raw" && !info.Closed {
-		historyReq.TailCompactions = 1
-	}
-	history, historyErr := handle.History(r.Context(), historyReq)
-	hasHistory := historyErr == nil && history != nil
-	if historyErr != nil && !errors.Is(historyErr, worker.ErrHistoryUnavailable) {
-		writeError(w, http.StatusInternalServerError, "internal", "reading session history: "+historyErr.Error())
-		return
-	}
-
-	state, stateErr := handle.State(r.Context())
-	if stateErr != nil {
-		writeSessionManagerError(w, stateErr)
-		return
-	}
-	running := workerPhaseHasLiveOutput(state.Phase)
-	if !hasHistory && !running {
-		writeError(w, http.StatusNotFound, "not_found", "session "+id+" has no live output")
-		return
-	}
-
-	w.Header().Set("Content-Type", "text/event-stream")
-	w.Header().Set("Cache-Control", "no-cache")
-	w.Header().Set("Connection", "keep-alive")
-	if info.State != "" {
-		w.Header().Set("GC-Session-State", string(info.State))
-	}
-	if !running {
-		w.Header().Set("GC-Session-Status", "stopped")
-	}
-	w.WriteHeader(http.StatusOK)
-	if err := http.NewResponseController(w).Flush(); err != nil {
-		_ = err
-	}
-
-	ctx := r.Context()
-	send := sse.Sender(func(msg sse.Message) error {
-		switch data := msg.Data.(type) {
-		case SessionStreamRawMessageEvent:
-			payload, err := json.Marshal(data)
-			if err != nil {
-				return err
-			}
-			writeSSE(w, "message", msg.ID, payload)
-		case SessionStreamMessageEvent:
-			payload, err := json.Marshal(data)
-			if err != nil {
-				return err
-			}
-			writeSSE(w, "turn", msg.ID, payload)
-		case SessionActivityEvent:
-			payload, err := json.Marshal(data)
-			if err != nil {
-				return err
-			}
-			writeSSE(w, "activity", msg.ID, payload)
-		case runtime.PendingInteraction:
-			payload, err := json.Marshal(data)
-			if err != nil {
-				return err
-			}
-			writeSSE(w, "pending", msg.ID, payload)
-		case HeartbeatEvent:
-			writeSSEComment(w)
-		}
-		return nil
-	})
-	format := r.URL.Query().Get("format")
-	if info.Closed {
-		if format == "raw" {
-			s.emitClosedSessionSnapshotRaw(w, info, history)
-		} else {
-			s.emitClosedSessionSnapshot(w, info, history)
-		}
-		return
-	}
-	switch {
-	case hasHistory:
-		if format == "raw" {
-			s.streamSessionTranscriptHistoryRaw(ctx, w, info, handle, history, historyReq)
-		} else {
-			s.streamSessionTranscriptHistory(ctx, w, info, handle, history)
-		}
-	case format == "raw":
-		// No log file yet. If the session is running, poll tmux pane content
-		// and wrap it as a fake raw JSONL assistant message so MC's existing
-		// rendering pipeline shows terminal output (e.g. OAuth prompts).
-		if running {
-			s.streamSessionPeekRaw(ctx, send, info)
-		} else {
-			data, _ := json.Marshal(sessionRawTranscriptResponse{
-				ID:       info.ID,
-				Template: info.Template,
-				Format:   "raw",
-				Messages: []json.RawMessage{},
-			})
-			writeSSE(w, "message", 1, data)
-		}
-		return
-	default:
-		s.streamSessionPeek(ctx, send, info)
-	}
-}
-
-func workerPhaseHasLiveOutput(phase worker.Phase) bool {
-	switch phase {
-	case worker.PhaseStarting, worker.PhaseReady, worker.PhaseBusy, worker.PhaseBlocked, worker.PhaseStopping:
-		return true
-	default:
-		return false
-	}
-}
-
-func (s *Server) emitClosedSessionSnapshot(w http.ResponseWriter, info session.Info, history *worker.HistorySnapshot) {
-	if history == nil {
-		return
-	}
-	turns, _ := historySnapshotTurns(history)
-	if len(turns) == 0 {
-		return
-	}
-
-	data, err := json.Marshal(sessionTranscriptResponse{
-		ID:       info.ID,
-		Template: info.Template,
-		Format:   "conversation",
-		Turns:    turns,
-	})
-	if err != nil {
-		return
-	}
-	writeSSE(w, "turn", 1, data)
-	// Closed session is definitionally idle.
-	actData, _ := json.Marshal(map[string]string{"activity": "idle"})
-	writeSSE(w, "activity", 2, actData)
-}
-
-func (s *Server) emitClosedSessionSnapshotRaw(w http.ResponseWriter, info session.Info, history *worker.HistorySnapshot) {
-	if history == nil {
-		return
-	}
-	rawMessages, _ := historySnapshotRawMessages(history)
-	if len(rawMessages) == 0 {
-		return
-	}
-
-	data, err := json.Marshal(sessionRawTranscriptResponse{
-		ID:       info.ID,
-		Template: info.Template,
-		Format:   "raw",
-		Messages: rawMessages,
-	})
-	if err != nil {
-		return
-	}
-	writeSSE(w, "message", 1, data)
-	actData, _ := json.Marshal(map[string]string{"activity": "idle"})
-	writeSSE(w, "activity", 2, actData)
-}
-
-func (s *Server) streamSessionTranscriptHistoryRaw(ctx context.Context, w http.ResponseWriter, info session.Info, handle interface {
-	worker.HistoryHandle
-	worker.InteractionHandle
-}, initial *worker.HistorySnapshot, req worker.HistoryRequest,
-) {
-	poll := time.NewTicker(outputStreamPollInterval)
-	defer poll.Stop()
-	keepalive := time.NewTicker(sseKeepalive)
-	defer keepalive.Stop()
-
-	var lastSentID string
-	var seq uint64
-	var lastActivity string
-	var lastPendingID string
-	lastProgress := time.Now()
-	sentIDs := make(map[string]struct{})
-	currentActivity := historySnapshotActivity(initial)
-
-	emitSnapshot := func(snapshot *worker.HistorySnapshot) {
-		if snapshot == nil {
-			return
-		}
-		currentActivity = historySnapshotActivity(snapshot)
-		rawMessages, ids := historySnapshotRawMessages(snapshot)
-		if len(rawMessages) > 0 {
-			var toSend []json.RawMessage
-			if lastSentID == "" {
-				toSend = rawMessages
-			} else {
-				found := false
-				for i, id := range ids {
-					if id == lastSentID {
-						toSend = rawMessages[i+1:]
-						found = true
-						break
-					}
-				}
-				if !found {
-					log.Printf("session stream raw: cursor %s lost, emitting only new messages", lastSentID)
-					for i, id := range ids {
-						if _, seen := sentIDs[id]; !seen {
-							toSend = append(toSend, rawMessages[i])
-						}
-					}
-				}
-			}
-			if len(toSend) > 0 {
-				seq++
-				data, err := json.Marshal(sessionRawTranscriptResponse{
-					ID:       info.ID,
-					Template: info.Template,
-					Format:   "raw",
-					Messages: toSend,
-				})
-				if err == nil {
-					writeSSE(w, "message", seq, data)
-					lastProgress = time.Now()
-					lastPendingID = ""
-				}
-			}
-			lastSentID = ids[len(ids)-1]
-			for _, id := range ids {
-				sentIDs[id] = struct{}{}
-			}
-		}
-		if currentActivity != "" && currentActivity != lastActivity {
-			lastActivity = currentActivity
-			seq++
-			actData, _ := json.Marshal(map[string]string{"activity": currentActivity})
-			writeSSE(w, "activity", seq, actData)
-			lastProgress = time.Now()
-		}
-	}
-
-	emitPending := func() {
-		if time.Since(lastProgress) < 5*time.Second {
-			return
-		}
-		pending, err := handle.Pending(ctx)
-		if err != nil || pending == nil {
-			if lastPendingID != "" {
-				lastPendingID = ""
-				activity := currentActivity
-				if activity == "" {
-					activity = "in-turn"
-				}
-				seq++
-				actData, _ := json.Marshal(map[string]string{"activity": activity})
-				writeSSE(w, "activity", seq, actData)
-			}
-			return
-		}
-		if pending.RequestID == lastPendingID {
-			return
-		}
-		lastPendingID = pending.RequestID
-		seq++
-		pendingData, _ := json.Marshal(pending)
-		writeSSE(w, "pending", seq, pendingData)
-	}
-
-	emitSnapshot(initial)
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-poll.C:
-			snapshot, err := handle.History(ctx, req)
-			switch {
-			case err == nil:
-				emitSnapshot(snapshot)
-			case errors.Is(err, worker.ErrHistoryUnavailable):
-			default:
-				log.Printf("session stream raw: history reload failed for %s: %v", info.ID, err)
-			}
-			emitPending()
-		case <-keepalive.C:
-			writeSSEComment(w)
-		}
-	}
-}
-
-func (s *Server) streamSessionTranscriptHistory(ctx context.Context, w http.ResponseWriter, info session.Info, handle worker.HistoryHandle, initial *worker.HistorySnapshot) {
-	poll := time.NewTicker(outputStreamPollInterval)
-	defer poll.Stop()
-	keepalive := time.NewTicker(sseKeepalive)
-	defer keepalive.Stop()
-
-	var lastSentID string
-	var seq uint64
-	var lastActivity string
-	sentIDs := make(map[string]struct{})
-
-	emitSnapshot := func(snapshot *worker.HistorySnapshot) {
-		if snapshot == nil {
-			return
-		}
-		turns, ids := historySnapshotTurns(snapshot)
-		if len(turns) > 0 {
-			var toSend []outputTurn
-			if lastSentID == "" {
-				toSend = turns
-			} else {
-				found := false
-				for i, id := range ids {
-					if id == lastSentID {
-						toSend = turns[i+1:]
-						found = true
-						break
-					}
-				}
-				if !found {
-					log.Printf("session stream: cursor %s lost, emitting only new turns", lastSentID)
-					for i, id := range ids {
-						if _, seen := sentIDs[id]; !seen {
-							toSend = append(toSend, turns[i])
-						}
-					}
-				}
-			}
-			if len(toSend) > 0 {
-				seq++
-				data, err := json.Marshal(sessionTranscriptResponse{
-					ID:       info.ID,
-					Template: info.Template,
-					Format:   "conversation",
-					Turns:    toSend,
-				})
-				if err == nil {
-					writeSSE(w, "turn", seq, data)
-				}
-			}
-			lastSentID = ids[len(ids)-1]
-			for _, id := range ids {
-				sentIDs[id] = struct{}{}
-			}
-		}
-		activity := historySnapshotActivity(snapshot)
-		if activity != "" && activity != lastActivity {
-			lastActivity = activity
-			seq++
-			actData, _ := json.Marshal(map[string]string{"activity": activity})
-			writeSSE(w, "activity", seq, actData)
-		}
-	}
-
-	emitSnapshot(initial)
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-poll.C:
-			snapshot, err := handle.History(ctx, worker.HistoryRequest{})
-			switch {
-			case err == nil:
-				emitSnapshot(snapshot)
-			case errors.Is(err, worker.ErrHistoryUnavailable):
-			default:
-				log.Printf("session stream: history reload failed for %s: %v", info.ID, err)
-			}
-		case <-keepalive.C:
-			writeSSEComment(w)
-		}
-	}
-}
-
-func (s *Server) streamSessionTranscriptLogRaw(ctx context.Context, send sse.Sender, info session.Info, logPath string) {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-	send = cancelOnSendError(send, cancel)
-
-	lw := newLogFileWatcher(logPath)
-	defer lw.Close()
-
-	var lastSize int64
-	var lastSentUUID string
-	var seq int
-	var lastActivity string
-	sentUUIDs := make(map[string]struct{})
-	lw.onReset = func() { lastSize = 0; lastActivity = "" }
-
-	readAndEmit := func() {
-		stat, err := os.Stat(logPath)
-		if err != nil {
-			return
-		}
-		if stat.Size() == lastSize {
-			return
-		}
-
-		// Use tail=1 (last compaction segment) to limit parsing scope,
-		// consistent with the non-raw streaming path.
-		factory, err := s.workerFactory(s.state.CityBeadStore())
-		if err != nil {
-			return
-		}
-		transcript, err := factory.ReadTranscript(worker.TranscriptRequest{
-			Provider:        info.Provider,
-			TranscriptPath:  logPath,
-			TailCompactions: 1,
-			Raw:             true,
-		})
-		if err != nil {
-			return
-		}
-		sess := transcript.Session
-		lastSize = stat.Size()
-
-		// Compute activity early (used after message emission).
-		activity := worker.InferTranscriptActivity(sess.Messages)
-
-		// Keep raw bytes end-to-end. Previously we Unmarshaled entry.Raw
-		// into `any` and remarshaled in wrapRawFrames — that round-trip
-		// loses int64 precision above 2^53 (tool-call IDs, nanosecond
-		// timestamps) and does not preserve map-key order. Provider-native
-		// frames must ship byte-faithful; we use json.RawMessage so the
-		// wire output matches what the provider wrote verbatim.
-		rawBytes := make([]json.RawMessage, 0, len(sess.Messages))
-		uuids := make([]string, 0, len(sess.Messages))
-		for _, entry := range sess.Messages {
-			if len(entry.Raw) == 0 {
-				continue
-			}
-			// Validate that the bytes are well-formed JSON; skip malformed
-			// frames the same way the previous Unmarshal branch did.
-			if !json.Valid(entry.Raw) {
-				continue
-			}
-			rawBytes = append(rawBytes, entry.Raw)
-			uuids = append(uuids, entry.UUID)
-		}
-
-		// Emit messages if there are new ones.
-		if len(rawBytes) > 0 {
-			var toSend []json.RawMessage
-
-			if lastSentUUID == "" {
-				// First emission: send everything.
-				toSend = rawBytes
-			} else {
-				found := false
-				for i, uuid := range uuids {
-					if uuid == lastSentUUID {
-						toSend = rawBytes[i+1:]
-						found = true
-						break
-					}
-				}
-				if !found {
-					// Cursor lost (DAG rewrite, compaction). Instead of
-					// re-syncing from the beginning (which causes duplicate/
-					// out-of-order messages on the client), emit only messages
-					// we haven't previously sent.
-					log.Printf("session stream raw: cursor %s lost, emitting only new messages", lastSentUUID)
-					for i, uuid := range uuids {
-						if _, seen := sentUUIDs[uuid]; !seen {
-							toSend = append(toSend, rawBytes[i])
-						}
-					}
-				}
-			}
-
-			if len(toSend) > 0 {
-				seq++
-				_ = send(sse.Message{ID: seq, Data: SessionStreamRawMessageEvent{
-					ID:       info.ID,
-					Template: info.Template,
-					Provider: info.Provider,
-					Format:   "raw",
-					Messages: wrapRawFrameBytes(toSend),
-				}})
-			}
-
-			// Track all current UUIDs so cursor-lost can filter correctly.
-			lastSentUUID = uuids[len(uuids)-1]
-			for _, uuid := range uuids {
-				sentUUIDs[uuid] = struct{}{}
-			}
-		}
-
-		// Emit activity after content so clients receive data before state change.
-		if activity != "" && activity != lastActivity {
-			lastActivity = activity
-			seq++
-			_ = send(sse.Message{ID: seq, Data: SessionActivityEvent{Activity: activity}})
-		}
-	}
-
-	// Stall detection: when the log hasn't grown for 5s, check the tmux
-	// pane for a tool approval prompt. If found, emit a "pending" SSE event
-	// so the UI can show the approval panel.
-	var lastPendingID string
-	onStall := func() {
-		sp := s.state.SessionProvider()
-		ip, ok := sp.(runtime.InteractionProvider)
-		if !ok {
-			return
-		}
-		pending, err := ip.Pending(info.SessionName)
-		if err != nil || pending == nil {
-			if lastPendingID != "" {
-				// Approval cleared — emit activity update.
-				lastPendingID = ""
-				seq++
-				_ = send(sse.Message{ID: seq, Data: SessionActivityEvent{Activity: "in-turn"}})
-			}
-			return
-		}
-		if pending.RequestID == lastPendingID {
-			return // already emitted this approval
-		}
-		lastPendingID = pending.RequestID
-		seq++
-		_ = send(sse.Message{ID: seq, Data: *pending})
-	}
-
-	keepaliveTicker := time.NewTicker(sseKeepalive)
-	defer keepaliveTicker.Stop()
-	lw.Run(ctx, readAndEmit, func() {
-		_ = send.Data(HeartbeatEvent{Timestamp: time.Now().UTC().Format(time.RFC3339)})
-	}, RunOpts{
-		OnStall:      onStall,
-		StallTimeout: 5 * time.Second,
-	})
-}
-
-func (s *Server) streamSessionTranscriptLog(ctx context.Context, send sse.Sender, info session.Info, logPath string) {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-	send = cancelOnSendError(send, cancel)
-
-	lw := newLogFileWatcher(logPath)
-	defer lw.Close()
-
-	var lastSize int64
-	var lastSentUUID string
-	var seq int
-	var lastActivity string
-	sentUUIDs := make(map[string]struct{})
-	lw.onReset = func() { lastSize = 0; lastActivity = "" }
-
-	readAndEmit := func() {
-		stat, err := os.Stat(logPath)
-		if err != nil {
-			return
-		}
-		if stat.Size() == lastSize {
-			return
-		}
-
-		factory, err := s.workerFactory(s.state.CityBeadStore())
-		if err != nil {
-			return
-		}
-		transcript, err := factory.ReadTranscript(worker.TranscriptRequest{
-			Provider:       info.Provider,
-			TranscriptPath: logPath,
-		})
-		if err != nil {
-			return
-		}
-		sess := transcript.Session
-		lastSize = stat.Size()
-
-		// Compute activity early (used after turn emission).
-		activity := worker.InferTranscriptActivity(sess.Messages)
-
-		turns := make([]outputTurn, 0, len(sess.Messages))
-		uuids := make([]string, 0, len(sess.Messages))
-		for _, entry := range sess.Messages {
-			turn := entryToTurn(entry)
-			if turn.Text == "" {
-				continue
-			}
-			turns = append(turns, turn)
-			uuids = append(uuids, entry.UUID)
-		}
-
-		// Emit turns if there are new ones.
-		if len(turns) > 0 {
-			var toSend []outputTurn
-
-			if lastSentUUID == "" {
-				// First emission: send everything.
-				toSend = turns
-			} else {
-				found := false
-				for i, uuid := range uuids {
-					if uuid == lastSentUUID {
-						toSend = turns[i+1:]
-						found = true
-						break
-					}
-				}
-				if !found {
-					// Cursor lost (DAG rewrite, compaction). Instead of
-					// re-syncing from the beginning (which causes duplicate/
-					// out-of-order messages on the client), emit only turns
-					// we haven't previously sent.
-					log.Printf("session stream: cursor %s lost, emitting only new turns", lastSentUUID)
-					for i, uuid := range uuids {
-						if _, seen := sentUUIDs[uuid]; !seen {
-							toSend = append(toSend, turns[i])
-						}
-					}
-				}
-			}
-
-			if len(toSend) > 0 {
-				seq++
-				_ = send(sse.Message{ID: seq, Data: SessionStreamMessageEvent{
-					ID:       info.ID,
-					Template: info.Template,
-					Provider: info.Provider,
-					Format:   "conversation",
-					Turns:    toSend,
-				}})
-			}
-
-			// Track all current UUIDs so cursor-lost can filter correctly.
-			lastSentUUID = uuids[len(uuids)-1]
-			for _, uuid := range uuids {
-				sentUUIDs[uuid] = struct{}{}
-			}
-		}
-
-		// Emit activity after content so clients receive data before state change.
-		if activity != "" && activity != lastActivity {
-			lastActivity = activity
-			seq++
-			_ = send(sse.Message{ID: seq, Data: SessionActivityEvent{Activity: activity}})
-		}
-	}
-
-	lw.Run(ctx, readAndEmit, func() {
-		_ = send.Data(HeartbeatEvent{Timestamp: time.Now().UTC().Format(time.RFC3339)})
-	})
-}
-
-// streamSessionPeekRaw polls tmux pane content and wraps it as format=raw
-// messages so MC's JSONL rendering pipeline can display terminal output
-// (e.g. OAuth prompts, startup screens) when no transcript log exists yet.
-func (s *Server) streamSessionPeekRaw(ctx context.Context, send sse.Sender, info session.Info) {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-	send = cancelOnSendError(send, cancel)
-	handle, err := s.workerHandleForSession(s.state.CityBeadStore(), info.ID)
-	if err != nil {
-		return
-	}
-	poll := time.NewTicker(outputStreamPollInterval)
-	defer poll.Stop()
-	keepalive := time.NewTicker(sseKeepalive)
-	defer keepalive.Stop()
-
-	var lastOutput string
-	var seq int
-
-	var lastPeekPendingID string
-
-	emitPeek := func() {
-		output, err := handle.Peek(ctx, 100)
-		if errors.Is(err, session.ErrSessionInactive) {
-			return
-		}
-		if err != nil || output == lastOutput {
-			return
-		}
-		lastOutput = output
-		seq++
-
-		if output == "" {
-			return
-		}
-
-		// Wrap as a fake assistant message in raw JSONL format so MC's
-		// translate_transcript_response handles it like a normal transcript.
-		fakeMsg := map[string]any{
-			"role": "assistant",
-			"content": []map[string]string{
-				{"type": "text", "text": output},
-			},
-		}
-		_ = send(sse.Message{ID: seq, Data: SessionStreamRawMessageEvent{
-			ID:       info.ID,
-			Template: info.Template,
-			Provider: info.Provider,
-			Format:   "raw",
-			Messages: wrapRawFrames([]any{fakeMsg}),
-		}})
-
-		// Check for approval prompts in the pane output we already have.
-		pending, pErr := handle.Pending(ctx)
-		if pErr == nil && pending != nil && pending.RequestID != lastPeekPendingID {
-			lastPeekPendingID = pending.RequestID
-			seq++
-			_ = send(sse.Message{ID: seq, Data: *pending})
-		} else if pending == nil && lastPeekPendingID != "" {
-			lastPeekPendingID = ""
-		}
-	}
-
-	emitPeek()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-poll.C:
-			emitPeek()
-		case <-keepalive.C:
-			_ = send.Data(HeartbeatEvent{Timestamp: time.Now().UTC().Format(time.RFC3339)})
-		}
-	}
-}
-
-func (s *Server) streamSessionPeek(ctx context.Context, send sse.Sender, info session.Info) {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-	send = cancelOnSendError(send, cancel)
-	handle, err := s.workerHandleForSession(s.state.CityBeadStore(), info.ID)
-	if err != nil {
-		return
-	}
-	poll := time.NewTicker(outputStreamPollInterval)
-	defer poll.Stop()
-	keepalive := time.NewTicker(sseKeepalive)
-	defer keepalive.Stop()
-
-	var lastOutput string
-	var seq int
-
-	emitPeek := func() {
-		output, err := handle.Peek(ctx, 100)
-		if errors.Is(err, session.ErrSessionInactive) {
-			return
-		}
-		if err != nil || output == lastOutput {
-			return
-		}
-		lastOutput = output
-		seq++
-
-		turns := []outputTurn{}
-		if output != "" {
-			turns = append(turns, outputTurn{Role: "output", Text: output})
-		}
-		_ = send(sse.Message{ID: seq, Data: SessionStreamMessageEvent{
-			ID:       info.ID,
-			Template: info.Template,
-			Provider: info.Provider,
-			Format:   "text",
-			Turns:    turns,
-		}})
-	}
-
-	emitPeek()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-poll.C:
-			emitPeek()
-		case <-keepalive.C:
-			_ = send.Data(HeartbeatEvent{Timestamp: time.Now().UTC().Format(time.RFC3339)})
-		}
-	}
 }
